+++ generators/app/index.js
@@ -2,8 +2,39 @@
 
 const Generator = require('yeoman-generator');
 const app = require('../../lib/generators/app');
+const optionOrPrompt = require('../../lib/utils').yeomanOptionOrPrompt;
 
 module.exports = class extends Generator {
+  constructor(args, opts) {
+    super(args, opts);
+
+    // This method adds support for a `--use-cli-defaults` flag
+    this.option('useDefaults', {
+      name: 'use-cli-defaults',
+      description: 'Use default values if not provided (will not prompt)',
+      hide: false
+    });
+
+    // This method adds support for a `--run-gradle-build` flag
+    this.option('runGradleBuild', {
+      name: 'run-gradle-build',
+      description: 'Run build gradle on generated project',
+      hide: false
+    });
+
+    app.prompts.forEach(prompt =>
+      this.option(prompt.name, {
+        name: prompt.cliName,
+        description: prompt.message,
+        type: String,
+        default: (this.options.useCliDefaults || this.options.help) ? prompt.default : undefined,
+        hide: false
+      })
+    );
+
+    this._optionOrPrompt = optionOrPrompt;
+  }
+
   initializing() {
     this.props = {};
   }
@@ -15,7 +46,7 @@ module.exports = class extends Generator {
   prompting() {
     const self = this;
     const done = self.async();
-    return this.prompt(app.prompts).then(function responses(props) {
+    return this._optionOrPrompt(app.prompts).then(function responses(props) {
       app.onResponses(self, props, done);
     });
   }
@@ -29,7 +60,11 @@ module.exports = class extends Generator {
   }
 
   end() {
-    // this.spawnCommand('./gradlew', ['build']);
+    if (this.options.runGradleBuild) {
+      this.spawnCommand('./gradlew', ['build']);
+    } else {
+      this.log('Skipping run gradle build (no --runGradleBuild flag set)');
+    }
   }
 
 };